+++ packages/blob-stores/fileSystemBlobStore.js
@@ -73,7 +73,7 @@ class FileSystemBlobStore extends BlobStoreCustom {
     } else if (this.storeSize === this.SIZES.Large) {
       this._folderCounter = (getRandomInt(STORE_SUBFOLDER_COUNT) + 1) * (getRandomInt(STORE_SUBFOLDER_COUNT) + 1)
     }
-    if ((this.storeSize === this.SIZES.Simple) && (this.LUCount>0)) {
+    if ((this.storeSize === this.SIZES.Simple) && (this.LUCount > 0)) {
       throw new Error(`BLOB Store '${this.name}': LUCount can not be set for 'Simple' store`)
     }
   }
@@ -100,7 +100,7 @@ class FileSystemBlobStore extends BlobStoreCustom {
       if (fs.existsSync(fn)) fs.unlinkSync(fn)
       throw e
     }
-    const origFn = request.fileName
+    const origFn = request.fileName || 'no-file-name.bin'
     const ct = mime.contentType(path.extname(origFn)) || 'application/octet-stream'
     const newMD5 = nhashFile(fn, 'MD5')
     return {
@@ -183,6 +183,7 @@ class FileSystemBlobStore extends BlobStoreCustom {
         // }
         resp.writeEnd(filePath)
       }
+      return true
     } else {
       return preventChangeRespOnError
         ? false